Setup for Dagger 2
This gives us a base dependency-injected project to start the app from. We have an activity which opens a fragment which has a string injected into it and displayed onscreen as a proof of concept. Modules have been setup to show how things like Retrofit and Room will be setup. Classes such as FeelingAPI, FeelingRepository and the MainViewModel have been added for the sake of being able to see how the structure will look in the future, they are not in use now.